+++ pkg/manifest/encoder.go
@@ -1,58 +1,10 @@
 package manifest
 
-import (
-	"bytes"
+import ( "bytes"
 	"io"
 
 	"gopkg.in/yaml.v3"
 )
-
-type Kind string
-
-const (
-	KindPipeline Kind = "pipeline"
-	KindSecret   Kind = "secret"
-)
-
-type Resource interface {
-	GetKind() Kind
-}
-
-type resource struct {
-	Kind         Kind                   `yaml:"kind" json:"kind"`
-	ResourceData map[string]interface{} `yaml:",inline" json:",inline"`
-}
-
-func (r *resource) GetKind() Kind {
-	return r.Kind
-}
-
-type Toleration struct {
-	Key, Operator, Value, Effect string
-}
-
-type Pipeline struct {
-	Kind         Kind                   `yaml:"kind" json:"kind"`
-	Type         string                 `yaml:"type,omitempty" json:"type,omitempty"`
-	Name         string                 `yaml:"name,omitempty" json:"name,omitempty"`
-	NodeSelector map[string]string      `yaml:"node_selector,omitempty" json:"node_selector,omitempty"`
-	Tolerations  []Toleration           `yaml:"tolerations,omitempty" json:"tolerations,omitempty"`
-	ResourceData map[string]interface{} `yaml:",inline" json:",inline"`
-}
-
-func (p *Pipeline) GetKind() Kind {
-	return KindPipeline
-}
-
-type Secret struct {
-	Kind         Kind                   `yaml:"kind" json:"kind"`
-	Name         string                 `yaml:"name,omitempty" json:"name,omitempty"`
-	ResourceData map[string]interface{} `yaml:",inline" json:",inline"`
-}
-
-func (p *Secret) GetKind() Kind {
-	return KindSecret
-}
 
 func Decode(data string) ([]Resource, error) {
 	buf := bytes.NewBufferString(data)
