+++ pkg/manifest/encoder_test.go
